removed the service interfaces
