+++ src/main/java/easy/mind/com/api/service/TestsQuestionsService.java
@@ -1,15 +1,35 @@
 package easy.mind.com.api.service;
 
-import easy.mind.com.api.DTO.TestsQuestionsDTO;
-import easy.mind.com.api.DTO.UserDTO;
 import easy.mind.com.api.entity.TestsQuestions;
+import easy.mind.com.api.repository.TestsQuestionsRepository;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface TestsQuestionsService {
+@Service
+@RequiredArgsConstructor
+public class TestsQuestionsService {
 
-    TestsQuestions create(TestsQuestions questions);
-    TestsQuestions getById(long id);
-    void delete(long id);
-    List<TestsQuestions> getAll();
+    private final TestsQuestionsRepository repository;
+
+    @Transactional
+    public TestsQuestions create(TestsQuestions questions) {
+        return repository.save(questions);
+    }
+
+    public TestsQuestions getById(long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Questions with id " + id + " not found"));
+    }
+
+    public void delete(long id) {
+        repository.delete(getById(id));
+    }
+
+    public List<TestsQuestions> getAll() {
+        return repository.findAll();
+    }
 }
